Refactor package structure

+++ attention_unet/training/main_training_w_dropout.py
@@ -8,19 +8,20 @@ from tensorflow.keras import optimizers, mixed_precision
 from tensorflow.keras.models import load_model
 from tensorflow.keras.utils import Progbar
 from tensorflow.keras.regularizers import l2
-from sklearn.metrics import f1_score
-from Losses.losses import focal_loss, ce_loss
-from Processing.transforms import Transforms
-from Model.unets import build_unet
-from Model.custom_layers import AttentionBlock, GroupNormalization, SpatialConcreteDropout, DropoutAttentionBlock
-from cfg.config import load_config
-from Dataloader.dataloader import get_dataset
 
+from sklearn.metrics import f1_score
+from attention_unet.losses.losses import focal_loss, ce_loss
+from attention_unet.processing.transforms import Transforms
+from attention_unet.models.unets import build_mc_dropout_unet
+from attention_unet.models.custom_layers import SpatialConcreteDropout, DropoutAttentionBlock, GroupNormalization
+from attention_unet.config.config import load_config
+from attention_unet.dataloader.dataloader import get_dataset
+
+# Configure GPU settings
 gpus = tf.config.experimental.list_physical_devices('GPU')
 for gpu in gpus:
     tf.config.experimental.set_memory_growth(gpu, True)
 
-os.environ['TF_CPP_MIN_LOG_LEVEL'] = '0'
 mixed_precision.set_global_policy('mixed_float16')
 
 
@@ -36,6 +37,7 @@ def read_last_best_val_f1(log_file_path):
     except FileNotFoundError:
         return float('-inf')
 
+
 def get_last_epoch(log_file_path):
     try:
         with open(log_file_path, 'r') as file:
@@ -44,55 +46,52 @@ def get_last_epoch(log_file_path):
         return 0
 
 
+def gather_from_replicas(strategy, value):
+    gathered_value = strategy.gather(value, axis=0)
+    return gathered_value
+
+
 def compute_f1_score(y_true, y_pred):
-    # Extract the center pixel from each image in the batch
-    batch_size, height, width, num_classes = y_true.shape
-    center_x, center_y = height // 2, width // 2
+    y_true = np.squeeze(np.argmax(y_true, axis=-1))
+    y_pred = np.squeeze(np.argmax(y_pred, axis=-1))
+    return f1_score(y_true, y_pred, average='macro')
 
-    y_true_center = y_true[:, center_x, center_y, :]
-    y_pred_center = y_pred[:, center_x, center_y, :]
-
-    y_true_flat = tf.argmax(y_true_center, axis=-1).numpy()
-    y_pred_flat = tf.argmax(y_pred_center, axis=-1).numpy()
-
-    return f1_score(y_true_flat, y_pred_flat, average='macro')
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "-c", "--config",
         type=Path,
-        help=r'Config file (YAML), see example in cfg/example_config.yaml.',
+        help=r'Config file (YAML), see example in attention_unet/config/example_config.yaml.',
         default=Path('configurations/configuration.yaml'))
     args = parser.parse_args()
     cfg = load_config(args.config)
-    model_prefix = 'dropout'
+    model_prefix = ''
+
 
     @tf.function(jit_compile=cfg['USE_XLA'])
     def compiled_train_step(batch, class_weights):
         with tf.GradientTape() as tape:
             x, y = batch
             probs = model(x, training=True)
-            if cfg['USE_FOCAL_LOSS']:
-                loss = focal_loss(y_true=y, y_pred=probs, alpha_weights=class_weights)
-            else:
-                loss = ce_loss(y_true=y, y_pred=probs, class_weight=class_weights)
+            loss = focal_loss(y_true=y, y_pred=probs, alpha_weights=class_weights)
             scaled_total_loss = optimizer.get_scaled_loss(loss)
         scaled_grads = tape.gradient(scaled_total_loss, model.trainable_weights)
         grads = optimizer.get_unscaled_gradients(scaled_grads)
-        return loss, grads, probs, y
+        return loss, grads
 
 
-    def train_step(batch=None, class_weights=None):
-        loss, grads, probs, y = strategy.run(compiled_train_step, args=(batch, class_weights))
+    def train_step(batch=None, class_weights=1):
+        loss, grads = strategy.run(compiled_train_step, args=(batch, class_weights))
         per_replica_loss = strategy.reduce(tf.distribute.ReduceOp.SUM, loss, axis=None)
         strategy.run(apply_gradients, args=(grads,))
-        return per_replica_loss, probs, y
+        return per_replica_loss
 
 
     @tf.function
     def apply_gradients(grads):
         optimizer.apply_gradients(zip(grads, model.trainable_weights))
+
 
     @tf.function(jit_compile=cfg['USE_XLA'])
     def compiled_validation_step(batch):
@@ -115,7 +114,6 @@ if __name__ == '__main__':
 
     with open(checkpoint_dir + 'paras.txt', 'w') as config_log:
         print(cfg, file=config_log)
-
     model_save_log_path = f'{checkpoint_dir}{model_prefix}_model_save_log.txt'
     best_val_f1 = read_last_best_val_f1(model_save_log_path)
     error_log_path = f'{checkpoint_dir}{model_prefix}_errors_{cfg["MODEL_NAME"][:-3]}.txt'
@@ -134,7 +132,7 @@ if __name__ == '__main__':
         scaling_factor = global_batch_size / cfg['BATCH_SIZE']
         adjusted_learning_rate = original_learning_rate * scaling_factor
 
-        global_steps = cfg['STEPS'] // cfg['BATCH_SIZE']
+        global_steps = cfg['STEPS'] // cfg['BATCH_SIZE'] // strategy.num_replicas_in_sync
         training_ds = get_dataset(
             cfg=cfg,
             repeat=True,
@@ -146,7 +144,7 @@ if __name__ == '__main__':
         distributed_training_ds = strategy.experimental_distribute_dataset(training_ds)
 
         global_val_batch_size = cfg['VAL_BATCH_SIZE'] * strategy.num_replicas_in_sync
-        global_val_steps = cfg['VAL_STEPS'] // cfg['VAL_BATCH_SIZE']
+        global_val_steps = cfg['VAL_STEPS'] // cfg['BATCH_SIZE'] // strategy.num_replicas_in_sync
         validation_ds = get_dataset(
             cfg=cfg,
             repeat=False,
@@ -170,14 +168,13 @@ if __name__ == '__main__':
                 compile=False,
                 custom_objects={
                     'DropoutAttentionBlock': DropoutAttentionBlock,
-                    'AttentionBlock': AttentionBlock,
-                    'SpatialConcreteDropout': SpatialConcreteDropout,
                     'GroupNormalization': GroupNormalization,
+                    'SpatialConcreteDropout': SpatialConcreteDropout,
                 }
             )
         else:
             tf.print('Building new model...')
-            model = build_unet(
+            model = build_mc_dropout_unet(
                 input_size=cfg['INPUT_SIZE'],
                 num_classes=cfg['OUT_CHANNELS'],
                 num_levels=cfg['N_MODEL_LEVELS'],
@@ -191,22 +188,22 @@ if __name__ == '__main__':
         model.compile()
         model.summary()
 
-    # Initialize CSV for p_logit values
-    with open(p_logit_log_path, 'w', newline='') as csvfile:
-        writer = csv.writer(csvfile)
-        header = ['Epoch'] + [layer.name for layer in model.layers if isinstance(layer, SpatialConcreteDropout)]
-        writer.writerow(header)
+    if not os.path.exists(p_logit_log_path):
+        with open(p_logit_log_path, 'w', newline='') as csvfile:
+            writer = csv.writer(csvfile)
+            header = ['Epoch'] + [layer.name for layer in model.layers if isinstance(layer, SpatialConcreteDropout)]
+            writer.writerow(header)
 
     for epoch in range(start_epoch, cfg['EPOCHS']):
         ds = iter(distributed_training_ds)
-        progbar_metrics = ['Epoch', 'focal_loss' if cfg['USE_FOCAL_LOSS'] else 'ce_loss', 'mean_p_logit']
+        progbar_metrics = ['Epoch', 'focal_loss' if cfg['USE_FOCAL_LOSS'] else 'ce_loss']
         train_pb = Progbar(global_steps, stateful_metrics=progbar_metrics)
         avg_loss = 0.0
         num_batches = 0
         epoch_p_logit_values = {layer.name: None for layer in model.layers if isinstance(layer, SpatialConcreteDropout)}
 
         for _, x, y in ds:
-            loss, probs, y_true = train_step((x, y), class_weights)
+            loss = train_step((x, y), class_weights)
             avg_loss += loss
             num_batches += 1
 
@@ -214,12 +211,9 @@ if __name__ == '__main__':
                 if isinstance(layer, SpatialConcreteDropout):
                     epoch_p_logit_values[layer.name] = layer.p_logit.numpy().mean()
 
-            mean_p_logit = np.mean([value for value in epoch_p_logit_values.values() if value is not None])
-
             progbar_values = [
                 ('Epoch', int(epoch) + 1),
                 ('focal_loss' if cfg['USE_FOCAL_LOSS'] else 'ce_loss', avg_loss / num_batches),
-                ('mean_p_logit', tf.sigmoid(mean_p_logit).numpy()),
             ]
 
             train_pb.update(num_batches, values=progbar_values)
@@ -229,18 +223,21 @@ if __name__ == '__main__':
         avg_val_loss = 0.0
         num_val_batches = 0
         val_f1_scores = []
+        y_true_gathered = []
+        y_probs_gathered = []
 
         for _, yx, yy in val_ds:
             val_loss, val_probs, y_true = validation_step((yx, yy))
+            center_x, center_y = cfg['INPUT_SIZE'][0] // 2, cfg['INPUT_SIZE'][1] // 2
             avg_val_loss += val_loss
             num_val_batches += 1
-            val_f1_scores.append(compute_f1_score(y_true, val_probs))
+            y_true_gathered.append(gather_from_replicas(strategy, y_true)[:, center_x, center_y, :])
+            y_probs_gathered.append(gather_from_replicas(strategy, val_probs)[:, center_x, center_y, :])
             val_pb.update(
                 num_val_batches,
                 values=[
                     ('Epoch', int(epoch) + 1),
                     ('val_loss', avg_val_loss / num_val_batches),
-                    ('val_f1_score', np.mean(val_f1_scores))
                 ]
             )
 
@@ -251,15 +248,19 @@ if __name__ == '__main__':
                                  isinstance(layer, SpatialConcreteDropout)]
             writer.writerow(row)
 
+        final_f1_score = compute_f1_score(np.array([y[0, :] for y in y_true_gathered]),
+                                          np.array([y[0, :] for y in y_probs_gathered]))
         with open(f'{checkpoint_dir}{model_prefix}_errors_{cfg["MODEL_NAME"][:-3]}.txt', 'a') as f:
             f.write(
-                f'{avg_loss.numpy() / num_batches}\t{avg_val_loss.numpy() / num_val_batches}\t'
-                f'{np.mean(val_f1_scores)}\n')
+                f'{avg_loss.numpy() / num_batches}\t'
+                f'{avg_val_loss.numpy() / num_val_batches}\t'
+                f'{final_f1_score}\n'
+            )
 
         model.save(checkpoint_model)
 
-        if np.mean(val_f1_scores) > best_val_f1:
-            best_val_f1 = np.mean(val_f1_scores)
+        if final_f1_score > best_val_f1:
+            best_val_f1 = final_f1_score
             model.save(best_model)
             with open(model_save_log_path, 'a') as save_log:
                 save_log.write(
